2006-02-20 Erik Hjortsberg  <erik@katastrof.nu>
    * src/components/ogre/SceneManagers/EmberPagingSceneManager/include/EmberPagingLandScapeData2D_HeightField.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/Makefile.am, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_HeightField.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_HeightFieldBlendNeighbor.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_HeightFieldN.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_HeightFieldNTC.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_HeightFieldRaw.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_HeightFieldRawTC.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_HeightFieldTC.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeData2D_Spline.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeOcclusionVisibilityData.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapePageManager.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/include/OgrePagingLandScapeRenderable.h, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/DRGNURBSSurface.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/EmberPagingLandScapeData2D_HeightField.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgreDebugRectangle2D.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2DManager.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_HeightField.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_HeightFieldBlendNeighbor.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_HeightFieldN.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_HeightFieldNTC.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_HeightFieldRaw.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_HeightFieldRawTC.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_HeightFieldTC.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D_Spline.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeOptions.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapePage.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapePageManager.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeRenderable.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeRenderableManager.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeSceneManager.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTextureManager.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTexture_InstantBaseTextureShadowed.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTexture_Splatting2Edit.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTexture_Splatting7Edit.cpp, src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTexture_SplattingShader.cpp  ,src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTile.cpp ,src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeTileManager.cpp: updated to latest version of the PLSM

+++ src/components/ogre/SceneManagers/EmberPagingSceneManager/src/OgrePagingLandScapeData2D.cpp
@@ -226,7 +226,7 @@ namespace Ogre
                                                     const uint z,
                                                     Real &modificationHeight)
     {  
-        const uint arraypos = z * mSize + x;
+        const uint arraypos = static_cast <uint> (z * mSize + x);
         assert (mHeightData && arraypos < mMaxArrayPos);
 
         const Real maxH = PagingLandScapeData2DManager::getSingleton ().getMaxHeight();
@@ -371,7 +371,7 @@ namespace Ogre
     void PagingLandScapeData2D::setHeight(const uint x, const uint z, 
 					                       const Real h)
     {       
-        const uint Pos = (z * mSize)+ x;
+        const uint Pos = static_cast <uint> ((z * mSize)+ x);
         assert  (mHeightData &&  mMaxArrayPos > Pos); 
         setHeight(x, z, Pos, h);
     }
